+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.6.0] - 2026-02-18
+
+### Mejorado
+- UX de generación con progreso por etapas.
+- Cancelación segura.
+- Mensajes de resultado mejorados.
 ## [1.5.2] - 2026-02-18
 
 ### Añadido